Refactor entity models to use Integer IDs
Changed all entity classes to use Integer instead of Long for IDs. Simplified Doctor, Patient, Consultation, and Specialty models by removing unused fields and updating constructors, getters, and setters. Updated Entity interface to use Integer for ID methods. Added toString methods for better debugging and logging.

+++ client_admin/src/main/java/org/example/model/entity/Entity.java
@@ -1,7 +1,7 @@
 package org.example.model.entity;
 
 public interface Entity {
-    Long getId();
-    void setId(Long id);
+    Integer getId();
+    void setId(Integer id);
 }
